backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-split-settings==1.2.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.1.0
python-dotenv==0.21.0
